The picture and description at the end of the article
The picture and description at the end of the article don't match. The original text reads "In this instance, the user would be blocked from accessing any cloud app on their trip to North Korea as Contoso has blocked access from that location" but the picture shows a sign in from an IP address in the United States.